database table finding column lat and long changed to use double to have more decimals

+++ database/migrations/2017_09_27_100122_create_findings_table.php
@@ -16,8 +16,8 @@ class CreateFindingsTable extends Migration
         Schema::create('findings', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->float('lat');
-            $table->float('long');
+            $table->DOUBLE('lat');
+            $table->DOUBLE('long');
             $table->timestamps();
         });
